% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_load}
\alias{batch_load}
\title{Load a large list of mixed files from a folder}
\usage{
batch_load(
  path,
  file_list,
  env = .GlobalEnv,
  haven_arg = list(),
  readxl_arg = list(),
  csv_arg = list(),
  fst_arg = list(),
  rds_arg = list(),
  readr = TRUE
)
}
\arguments{
\item{path}{character vector defining the directory containing files to load.}

\item{file_list}{character vector of files within directory to load.}

\item{env}{R environment to assign data to; default set to \code{.GlobalEnv}.}

\item{haven_arg, readxl_arg, csv_arg, fst_arg, rds_arg}{additional parameters passed to load functions depending on file extension.}

\item{readr}{boolean value to determine if \code{\link[readr]{read_csv}} is used in preference to \code{\link[utils]{read.csv}}.}
}
\value{
data-sets assigned to designated R environment.
}
\description{
Quickly load a large mixed set of files into R environment.
}
\details{
This is only useful if all the files can be loaded with similar parameters. If certain columns or rows needs to be dropped
during loading, then it may be best to load those separately. The use-case with the \{relay\} package is to provide a way to
quickly take data files from a bundle into the environment. Loaded data will be assigned to the global environment by default with their
names all in lower-case.

Files with the following extensions can all be loaded at the same time:
\enumerate{
    \item sas7bdat via \{haven\}
    \item xlsx via \{readxl\}
    \item csv via \{utils\} or \{readr\}
    \item rds via \{base\}
}

Please note: when loading more troublesome/complex data, it may be easier to use the appropriate load functions directly.
}
