% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_batons}
\alias{clear_batons}
\title{Cleanup batons}
\usage{
clear_batons(baton, loc = NULL, envir = .GlobalEnv, ...)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}.}

\item{loc}{file path to clear YAML files related to baton tracking.}

\item{envir}{environment to clear \emph{baton} objects from the active R session; default set to \code{.GlobalEnv}.}

\item{...}{parameters passed to \code{\link{list.files}}}
}
\description{
Clear batons from a directory (YAML files) and/or R environment.
}
\examples{
\dontrun{
clear_batons(loc = '/tmp', recursive = TRUE)
}
}
