% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_dir}
\alias{copy_dir}
\title{Copy directory and all sub-folders}
\usage{
copy_dir(from, to, tar, extra_flags, ...)
}
\arguments{
\item{from, to}{character vectors, defining file paths.}

\item{tar}{parameter passed to \code{\link[utils]{tar}}; may require assigning to \code{'tar'} manually if \code{Sys.getenv('tar')} has no value set.}

\item{extra_flags}{parameter passed to \code{\link[utils]{tar}}; defaults to \code{'-C'} for the \code{bundle_dir}.}

\item{...}{additional parameters passed to \code{\link[utils]{tar}}}
}
\description{
Bundles into a tar, then unpacks in a new location. Works with nested folders/files. May need to use \code{full.names} to
ensure all subfiles are located and moved.
}
