% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_files}
\alias{copy_files}
\title{Copy files to new location}
\usage{
copy_files(path, file_names, to, ...)
}
\arguments{
\item{path}{character vector defining directory with files of interest.}

\item{file_names}{character vector of file names within the directory.}

\item{to}{character vector defining the destination directory for the files listed.}

\item{...}{additional parameters passed to \code{\link{file.copy}}}
}
\description{
Helper function wrapped around \code{\link{file.copy}}.
}
\details{
Basic functionality is making it simple to collapse \code{\link{file.copy}} with \code{\link{file.path}} and
ensure the \code{copy.date} is kept to \code{TRUE}.
}
