% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_baton}
\alias{create_baton}
\title{Create a baton}
\usage{
create_baton(
  content = list(),
  loc = NULL,
  auto_assign = FALSE,
  envir = .GlobalEnv,
  bundled = FALSE,
  bundle_params = list()
)
}
\arguments{
\item{content}{A list of custom content that the baton should carry.}

\item{loc}{The location of the associated YAML file (defaults to a temporary location).}

\item{auto_assign}{Boolean value to determine if name assignment is to be automatic).}

\item{envir}{Environment where \code{auto_assign} should write to.}

\item{bundled}{Boolean value to determine if the baton should be created within a bundle; default is set to \code{FALSE}.}

\item{bundle_params}{List of parameters to pass to \code{\link{create_bundle}}.}
}
\value{
S3 class object.
}
\description{
\code{create_baton} is a constructor of an S3 class used for tracking workflows.
}
\details{
By default, the content of the \emph{baton} will be empty, only the metadata will be populated. The S3 object created will have an associated
YAML file made, in a temporary location if no set location provided to \code{loc}. There is also the ability to automatically assign the S3 object
to the R environment in case that is easier to remember.

The metadata of a bundle includes the following:
\enumerate{
   \item id: unique ID of baton based upon time stamp and random numbers.
   \item relay_start: time stamp when baton first created (matched ID).
   \item relay_finish: time the baton was last passed (will not be populated if in middle of pass).
   \item all_passes: time stamps for all completed passes.
   \item pass_complete: whether or not the baton completed its latest pass.
   \item passes_completed: the number of successful passes completed.
   \item location: where the baton YAML file has been saved.
   \item dropped: boolean value of whether the baton has been dropped, signalling and end to the relay.
}
}
\note{
Although some generics (\code{\link{summary.baton}}) are used in \{relay\}, most functions just check for the class and do not proceed unless
it is a \emph{baton}. This may be rewritten to use only generics or R6 classes but right now this hybrid approach was used for simplicity.
}
\examples{
\dontrun{
my_baton <- create_baton(loc = file.path('path', 'to', 'save', 'yaml'))
}
}
