% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_bundle}
\alias{create_bundle}
\title{Create a bundle}
\usage{
create_bundle(baton, dir, tree, tag = "_bundle-", mode, ...)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}.}

\item{dir}{character value for loaction to create the bundle on the file system.}

\item{tree}{character vector for sub directories; default has no sub-directories set.}

\item{tag}{character value for prefixed label to the bundled folder.}

\item{mode}{passed to \code{\link{Sys.chmod}}.}

\item{...}{passed to \code{\link{dir.create}}.}
}
\value{
Character vector for the location of created bundle.
}
\description{
\code{create_bundle} is a helper function to create a set of skeleton folders under a main bundle directory.
}
\details{
Creating a bundle can be done at the same time as \code{\link{create_baton}}. However, it can be used without a baton as well
to easily make a set of empty folders to populate later. When using as part of a \{relay\}, the primary purpose is to store information that
passes checks from batons.
}
\examples{
\dontrun{
create_bundle(my_baton, dir = '/path/to/relay/raw_bundle',
              tree = c('raw', 'processed', 'metadata', 'output', '/raw/sub-raw'))
}
}
