% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{drop_baton}
\alias{drop_baton}
\title{Drop a baton}
\usage{
drop_baton(baton)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}}
}
\value{
S3 class object.
}
\description{
Drop a baton to prevent it from being passed/grabbed.
}
\details{
Performed to lock the \emph{baton} from any further use of \code{\link{pass_baton}} or \code{\link{grab_baton}}. If performed in
error, can reset by manually editing the YAML file, assuming one has permissions to tinker with the files created!
}
