% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files_recursive}
\alias{list_files_recursive}
\title{List all files under directory}
\usage{
list_files_recursive(path, recursive = TRUE, basename_only = TRUE, ...)
}
\arguments{
\item{path}{character vector defining the top directory to search downwards.}

\item{recursive}{passed to \code{\link{list.files}}; default set to \code{TRUE}.}

\item{basename_only}{boolean value, determines if only the base file names are output (i.e. no file path components).}

\item{...}{parameters passed to \code{\link{list.files}}.}
}
\description{
Helper function to quickly list all the files in a directory, and its sub-directories.
}
\details{
This function may look very similar to \code{\link{list.files}} but the main difference is it will ensure
nested files have just the file names and not directories.
}
