% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.baton}
\alias{plot.baton}
\title{Plot baton metadata}
\usage{
\method{plot}{baton}(
  baton,
  relative_time = TRUE,
  separate = FALSE,
  include_logs = FALSE,
  x_label_length = 3,
  point_offset = 0.02,
  ...
)
}
\arguments{
\item{baton}{S3 object of class 'baton'.}

\item{relative_time}{Boolean, should time be relative to start or absolute (in seconds).}

\item{separate}{Boolean, determine if pass numbers should be on separate lines.}

\item{include_logs}{Boolean, include or exclude log times.}

\item{x_label_length}{Numeric value, for how many x-axis ticks to draw.}

\item{point_offset}{Numeric value, proportion for how far to offset points from lines.}

\item{...}{Additional parameters to plotting features (not in use yet...)}
}
\description{
Will parse the baton metadata and logbook to provide graphical aides in determining when
pass, grab, and logging operations occured.
}
\examples{
\dontrun{
library(relay)

test_baton <- create_baton()
write_logbook(test_baton, 'Test message 1')
test_baton <- pass_baton(test_baton)
write_logbook(test_baton, 'Test message 2')
test_baton <- grab_baton(test_baton)

plot(test_baton)
plot(test_baton, relative_time =  TRUE, separate = TRUE, include_logs = TRUE)
}
}
