% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick_copy}
\alias{quick_copy}
\title{Quickly copy files based upon pattern alone}
\usage{
quick_copy(from, to, pattern, recursive = FALSE, grepl_params = list(), ...)
}
\arguments{
\item{from, to}{character vectors, defining file paths.}

\item{pattern}{passed to \code{\link{grepl}}.}

\item{recursive}{passed to \code{\link{grepl}}.}

\item{grepl_params}{additional parameters passed to \code{\link{grepl}}.}

\item{...}{additional parameters passed to \code{\link{file.copy}}.}
}
\description{
Use regular-expressions via \code{\link{grepl}} to quickly copy files to a new location.
}
\details{
Operates as a helper function that wraps around \code{\link{grepl}}, \code{\link{list.files}}, and \code{\link{file.copy}}.
For another helper function in \{relay\} that allows file copying, see \code{\link{copy_files}}.
}
