% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{read_logbook}
\alias{read_logbook}
\title{Read a baton's logbook}
\usage{
read_logbook(baton, loc = NULL, as.list = FALSE)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}}

\item{loc}{Location of YAML file that was saved from a \emph{baton}.}

\item{as.list}{Boolean value, determine if returns as vector or a list.}
}
\value{
S3 class object.
}
\description{
Read log information from a baton's logbook without affecting the metadata or contents.
}
\seealso{
\code{\link{write_logbook}}
}
