% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata of baton}
\usage{
read_metadata(baton, loc = NULL, subset = NULL)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}.}

\item{loc}{file path to clear YAML files related to baton tracking.}

\item{subset}{a integer or character vector of metadata elements to return (defaults to ALL elements). Elements include: id, relay_start, relay_finish, all_passes, pass_complete, passes_completed, location, and dropped.}
}
\description{
Read the metadata content of a baton, either in the environment, or from a file source. This has some functional overlaps with \code{summary} but
is specific to the metadata section. This function will not alter the metadata content, and is a read-only operation. If you want to read log information
try \code{\link{read_logbook}}.
}
\examples{
\dontrun{

batons_loc <- locate_batons('/location/where/bundlesorbatons/live', recursive = TRUE, full.name = TRUE)

# Grab metadata of interest to determine which to load
baton_meta_pass <- purrr::map_lgl(batons_loc, ~unlist(read_metadata(loc = ., subset = c('pass_complete'))))
baton_meta_finish <- purrr::map_chr(batons_loc, ~unlist(read_metadata(loc = ., subset = c('relay_finish'))))

# Equivalent in base R
baton_meta_pass <- vapply(batons_loc, function(x) unlist(read_metadata(loc = x, subset = c('pass_complete'))), logical(1), USE.NAMES = F)
baton_meta_finish <- vapply(batons_loc, function(x) unlist(read_metadata(loc = x, subset = c('relay_finish'))), character(1), USE.NAMES = F)

# Determine which baton to keep!
select_baton <- batons_loc[which(baton_meta_finish == max(baton_meta_finish) & baton_meta_pass)]

}
}
