% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{relocate_baton}
\alias{relocate_baton}
\title{Relocate a baton}
\usage{
relocate_baton(baton, loc, silent = FALSE)
}
\value{
S3 class object.
}
\description{
Internal function used to relocate a baton to a new location.
}
