% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{set_referee}
\alias{set_referee}
\title{Set a baton's referee}
\usage{
set_referee(
  baton,
  threshold = getOption("relay_referee", default = "TRACE"),
  suppressWarnings = TRUE,
  autoassign = TRUE,
  envir = .GlobalEnv,
  ...
)
}
\arguments{
\item{baton}{R object of S3 class, created by \code{\link{create_baton}}.}

\item{threshold}{character value for the minimum threshold for logging to occur (e.g. 'TRACE', 'DEBUG', 'MESSAGE', 'WARNING', 'ERROR').}

\item{suppressWarnings}{boolean value to determine if warning messages upon YAML write are ignored.}

\item{autoassign}{boolean value to determine if the passed baton is also refreshed automatically. Recommended as TRUE to avoid having to do manual assignment to provided baton.}

\item{envir}{Environment where baton exists, default set to .GlobalEnv, only needed when autoassign is TRUE. If deploying on RStudio Connect, may require using \code{knitr::knit_global()}.}

\item{...}{Additional parameters passed to \code{assign}}
}
\value{
S3 class object.
}
\description{
\code{set_referee} is a helper function to set the 'referee' of the baton; this controls what threshold of content is written to the logbook. The
updated metadata will occur both in the source YAML as well as the R object for paired consistency.
}
\details{
The default threshold is set to \code{'TRACE'} and can be overriden directly or by setting the global options via \code{options(relay_referee = 'TRACE')}.
}
\examples{
\dontrun{
my_baton <- create_baton()
set_referee(my_baton, 'MESSAGE')
write_logbook(my_baton, 'A super important message')
set_referee(my_baton, 'ERROR')
write_logbook(my_baton, 'This message ignored b/c the referee cares not')
read_logbook(my_baton)
}
}
