% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.baton_preview}
\alias{summary.baton_preview}
\title{Summarize batons}
\usage{
\method{summary}{baton_preview}(object, include_contents = TRUE, ...)
}
\arguments{
\item{object}{R object of S3 class, created by \code{\link{create_baton}}.}

\item{...}{non-functional, left for future development.}
}
\description{
Generic function to summarize and print information on \emph{baton} S3 class.
}
