% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_baton}
\alias{validate_baton}
\title{Validate baton}
\usage{
validate_baton(baton)
}
\description{
\code{validate_baton} is an internal function to \{relay\}. Determines if the \emph{baton} has the correct metadata compliant to being a true baton.
}
\details{
Basic validation checks used in constructing batons
}
